server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = "Currencies servive"
authors = ["Luferov Victor <lyferov@yandex.ru>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.5"
djangorestframework = "^3.13.1"
requests = "^2.28.0"
beautifulsoup4 = "^4.11.1"
openpyxl = "^3.0.10"
django-cors-headers = "^3.13.0"
django-filter = "^22.1"
PyYAML = "^6.0"
uritemplate = "^4.1.1"
drf-yasg = "^1.20.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
flake8-import-order = "^0.18.1"
flake8-docstrings = "^1.6.0"
flake8-builtins = "^1.5.3"
flake8-quotes = "^3.3.1"
flake8-comprehensions = "^3.10.0"
flake8-eradicate = "^1.2.0"
flake8-simplify = "^0.19.2"
flake8-spellcheck = "^0.28.0"
pep8-naming = "^0.13.0"
flake8-use-fstring = "^1.3"
flake8-annotations = "^2.9.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
